pyproject.toml
[project]
name = "langchain-middleware-solution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.3",
    "langchain-google-genai>=3.0.0",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
